% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimeVarManipulation.R
\name{AggregatedColumns}
\alias{AggregatedColumns}
\title{This function takes in a dataframe and a columntouse from the dataframe
and generates two new columns. Your data should be in long format
as the first new column generated will be a dummy variable column for
columnToUse. The second column is a sum of the columnToUse. Each of 
these values is passed to all other rows with the same UID.
Need to generalize to beyond just UID.}
\usage{
AggregatedColumns(DF, columnToUse, NewCol1, NewCol2)
}
\arguments{
\item{columnToUse}{}

\item{dataframe}{}

\item{NewColumn1}{}
}
\description{
This function takes in a dataframe and a columntouse from the dataframe
and generates two new columns. Your data should be in long format
as the first new column generated will be a dummy variable column for
columnToUse. The second column is a sum of the columnToUse. Each of 
these values is passed to all other rows with the same UID.
Need to generalize to beyond just UID.
}
