% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimeVarManipulation.R
\name{Crime_type_dummies}
\alias{Crime_type_dummies}
\title{Function to create dummy variable based on a list of strings}
\usage{
Crime_type_dummies(list, data = df, columnToUse, columnToCreate)
}
\arguments{
\item{list}{}

\item{columnToUse}{}

\item{dataframe}{}
}
\description{
Function to create dummy variable based on a list of strings
}
