% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimeVarManipulation.R
\name{gen_crime_type}
\alias{gen_crime_type}
\title{Function to use a list to alter crime_type column}
\usage{
gen_crime_type(list, data = df, columnToUse, columnToChange, varInput)
}
\arguments{
\item{list}{}

\item{columnToUse}{}

\item{columnToChange}{}

\item{dataframe}{}
}
\description{
Function to use a list to alter crime_type column
}
